+++ probables/countminsketch/countminsketch.py
@@ -55,7 +55,6 @@ class CountMinSketch(object):
         self.__int32_t_max = 2147483647
         self.__int64_t_min = -9223372036854775808
         self.__int64_t_max = 9223372036854775807
-        self.__uint64_t_max = 2 ** 64
 
         if is_valid_file(filepath):
             self.__load(filepath)
@@ -222,8 +221,8 @@ class CountMinSketch(object):
             res.append(self._bins[t_bin])
         self.__elements_added += num_els
 
-        if self.__elements_added > self.__int32_t_max:
-            self.__elements_added = self.__int32_t_max
+        if self.__elements_added > self.__int64_t_max:
+            self.__elements_added = self.__int64_t_max
         return self.__query_method(sorted(res))
 
     def remove(self, key, num_els=1):
@@ -257,8 +256,8 @@ class CountMinSketch(object):
                 self._bins[t_bin] = self.__int32_t_min
             res.append(self._bins[t_bin])
         self.__elements_added -= num_els
-        if self.__elements_added < self.__int32_t_min:
-            self.__elements_added = self.__int32_t_max
+        if self.__elements_added < self.__int64_t_min:
+            self.__elements_added = self.__int64_t_min
 
         return self.__query_method(sorted(res))