+++ tests/bloom_test.py
@@ -143,7 +143,8 @@ class TestBloomFilter(unittest.TestCase):
 
     def test_bf_jaccard_invalid_msg(self):
         ''' check invalid type in a jaccard index message '''
-        msg = 'The parameter second must be of type BloomFilter'
+        msg = ('The parameter second must be of type BloomFilter or '
+               'a BloomFilterOnDisk')
         blm = BloomFilter(est_elements=10, false_positive_rate=0.05)
         blm.add('this is a test')
         try:
@@ -159,7 +160,8 @@ class TestBloomFilter(unittest.TestCase):
 
     def test_bf_union_invalid_msg(self):
         ''' check invalid type in a union message '''
-        msg = 'The parameter second must be of type BloomFilter'
+        msg = ('The parameter second must be of type BloomFilter or '
+               'a BloomFilterOnDisk')
         blm = BloomFilter(est_elements=10, false_positive_rate=0.05)
         blm.add('this is a test')
         try:
@@ -175,7 +177,8 @@ class TestBloomFilter(unittest.TestCase):
 
     def test_bf_intersec_invalid_msg(self):
         ''' check invalid type in a intersection message '''
-        msg = 'The parameter second must be of type BloomFilter'
+        msg = ('The parameter second must be of type BloomFilter or '
+               'a BloomFilterOnDisk')
         blm = BloomFilter(est_elements=10, false_positive_rate=0.05)
         blm.add('this is a test')
         try:
@@ -555,4 +558,45 @@ class TestBloomFilterOnDisk(unittest.TestCase):
         for idx in range(blm.bloom_length):
             self.assertEqual(blm._get_element(idx), 0)
 
+        os.remove(filename)
+
+    def test_bfod_union_diff(self):
+        ''' make sure checking for different bloom filters on disk works union
+        '''
+        filename = 'tmp.blm'
+        blm = BloomFilterOnDisk(filename, est_elements=10, false_positive_rate=0.05)
+        blm.add('this is a test')
+        blm2 = BloomFilter(est_elements=10, false_positive_rate=0.05,
+                           hash_function=different_hash)
+
+        blm3 = blm.union(blm2)
+        self.assertEqual(blm3, None)
+        os.remove(filename)
+
+    def test_bfod_intersection_diff(self):
+        ''' make sure checking for different bloom filters on disk works
+            intersection
+        '''
+        filename = 'tmp.blm'
+        blm = BloomFilterOnDisk(filename, est_elements=10, false_positive_rate=0.05)
+        blm.add('this is a test')
+        blm2 = BloomFilter(est_elements=10, false_positive_rate=0.05,
+                           hash_function=different_hash)
+
+        blm3 = blm.intersection(blm2)
+        self.assertEqual(blm3, None)
+        os.remove(filename)
+
+    def test_bfod_jaccard_diff(self):
+        ''' make sure checking for different bloom filters on disk works
+            jaccard
+        '''
+        filename = 'tmp.blm'
+        blm = BloomFilterOnDisk(filename, est_elements=10, false_positive_rate=0.05)
+        blm.add('this is a test')
+        blm2 = BloomFilter(est_elements=10, false_positive_rate=0.05,
+                           hash_function=different_hash)
+
+        blm3 = blm.jaccard_index(blm2)
+        self.assertEqual(blm3, None)
         os.remove(filename)